FastAPI_full_course/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
mysql-connector-python==8.2.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
alembic==1.12.1
python-dotenv==1.0.0
aiofiles==23.2.1
pydantic-settings==2.1.0